Add comments explaining the {Server,Client}.tick() pattern
The caller is responsible for calling {Client,Server}.tick() at fixed
time intervals, moving the client simulator one time unit per call. For
Client we return a boolean indicating whether or not a packet is
generated in the most recently completed time unit. For Server the
return value, if not None, is the packet that has finished being
processed.